Stop partially assigning IDs to comments
Since IDs are attached to modules, it's not possible to assign
IDs to all comments anyway, without complicating the IDs significantly.

In particular this case is not really possible to handle while
giving each comment a unique ID:

  someFile.chpl:
    // comment 1
    module M { }
    // comment 2
    module N { }

The problem here is that because we're not storing the file paths in the
IDs (just module paths), the comments are outside of any module. We could
imagine giving them ids 0 and 1, but that'd need to be unique per file,
and then we'd have to have a sense of file paths in (some) IDs. The
argument at this moment is that it isn't worth it to do so but we might
revisit it later.

So, instead, a tool needing to work with comments (such as a
documentation tool) should create its own maps of Locations of Comments
after parsing based upon the vector in Builder::Result.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>