Skip building the docs in linux32 testing
This skips building chpldoc and running documentation checks on linux32.  The
machine we are running linux32 testing on has a version of Python 3 that is too
old for some of the newer Python packages we want to rely on for chpldoc.  Turn
testing there off until we've updated our linux32 machine

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>