Use plain histo-atomics-forall-opt.good to match comm=ofi, ugni.
We get one message set when compiling histo-atomics-forall-opt.chpl with
comm=none, another with comm=gasnet, and a third with comm=ofi and ugni
which support network atomics.  Since we already have comm-none.good and
comm-gasnet.good files for the first two cases, here adjust the plain
.good file (which has been unused) to match the comm=ofi and ugni case.
This also prepares for the future, since presumably once GASNet-EX is
fully integrated we will have network atomics with that as well, and
then the plain .good can be used for all the comm!=none cases.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>