Attempt to fix an easy case of zippered size mismatches
Issue #11428 captures a longstanding issue in which we
zipper a smaller thing and a larger one:

       var A3: [1..3] real;
       var A4: [1..4] real;

       forall (i,j) in zip(A3, A4) do ..

which is a longstanding challenge in Chapel's current
leader-follower iterator design.  However, it seems
that at some point we started permitting the opposite
case through as well:

       forall (i,j) in zip(A4, A3) do ...

which doesn't really make any sense since A3 doesn't have
3 items to yield, so should give an OOB error at some
point when trying to do what the leader says.  Yet... it
doesn't.

This is an attempt to fix that by putting bounds checking
into the follower iterator.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>