Add an error for zippered follower iterations over too-small ranges
This adds an error for zippered iterations in which the follower is a
range that does not have enough elements in it.  This is a case that
was already generating an error, but not a very clear one, and is a
stepping stone toward generating similar such errors for domains and
arrays.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>