Restore the default value of these arguments for the deprecated functions
The default values were removed when the new overload shared the same name to
avoid resolution errors when the default value was relied upon.  However, now
that the new function no longer shares the same name, we want users relying on
the default value to get the deprecation message rather than a resolution error.
So restore the default value to those arguments.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>