Add calls to lock in the deprecation message for the default value calls, too
Prior to renaming the methods, calls using the default value would not generate
the deprecation warning because we had ensured the version using the new enum
would be preferred in that case.  But, with the new versions of the methods
using a new name entirely, such calls should result in deprecation warnings, so
lock in that those occur.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>