Issue an error when a DR domain/array is following a larger leader
Nikhil showed a case on chat this week where—to my surprise—we weren't
generating OOB errors for having a DR array get indexed out of bounds
when it was following a larger leader array.  This is something that
we used to catch, but no longer seem to.  Here, I'm adding a check to
the DR domain follower code to catch such cases.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>