Fix bug with isConstAssignable and operator methods
Prior to this change, we were looking at the wrong argument of the assignment
operator when determining the assignment style of types with operator method
assignment functions.  Rectify that issue by adjusting the argument we check if
the assignment function is an operator method.

With this change, AtomicT can now use operator methods (there was a test that
was checking if it specifically was const assignable, and we were getting the
wrong answer because of this bug)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>