Make support for c_fn_ptr's automatically available
Currently, 'c_fn_ptr' is a type that's injected by the compiler and
therefore always available to Chapel programmers, unlike other C
interop types.  This is unfortunate, but more than I can currently
fix, so I filed issue #17994 against it instead.

When experimenting with 'c_fn_ptr' in response to a user query, I
noticed that the functions which prevent calls to it are defined in
CPtr.chpl, which means that they aren't automatically available.
Here, I've moved those functions—and any others specific to c_fn_ptr—
into ChapelBase.chpl to reflect that they are built-in / always
available.  Like so many things, this is essentially treating
ChapelBase as a kitchen sink (sorry).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>