Reflect review: use same AM progress function for recv and xmit waits.
While reviewing this PR we noted much similarity between the code to
handle completion events and progress in the main loop of the AM handler
and in the progress function the AM handler calls while waiting for its
own initiated operations to complete, when that's needed.  This was a
cleanup we knew we needed to do, but one which had been put off and/or
forgotten more than once.  Here, finally tackle it.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>