Add unique prefix to bundled hwloc symbols.
The bundled hwloc is currently v1, which can cause a version mismatch with third-party software such
as libopen-pal that requires the v2 version of hwloc. Adding a prefix to the bundled hwloc symbols
ensures that only software that uses the prefix will use the bundled hwloc.

This change involves adding a --with-hwloc-symbol-prefix configuration option to hwloc. Note that
hwloc/rename.h creates macros for the original hwloc symbol names so that neither hwloc nor the
software that uses it must be modified to use the new symbol names. For Chapel this includes the
runtime source as well as the third-party qthreads library. However, qthread configuration fails if
hwloc provides non-standard symbol names, requiring a change to qthread to specify
the --with-hwloc-symbol-prefix option when configuring it to use the bundled hwloc.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>