Disable hwloc in gasnet
GASNet auto-detects hwloc if it's available at build time. On some
machines a system hwloc is available and gasnet is detecting that and
trying to use it, but at link time our hwloc is getting used. We added a
prefix to our hwloc in 18040, which is causing name mismatches at link
time. We should probably point gasnet to our hwloc in the longer term,
but for now just disable gasnet's use of hwloc since we don't currently
rely on it.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>