Stop hiding printchplenv errors in the compiler
printchplenv is run from the compiler and previously any fatal errors
weren't visible to users. We hid stderr to avoid getting warnings into
the compiler output, but this made fatal errors tricky to debug.

To resolve this add a way to suppress warnings, use that, and stop
redirecting stderr so users will see the fatal error messages.

For 18021

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>