Add a simple check for code that can run on a GPU that triggers for stream
Add a check for only calling FAST_AND_LOCAL primitives in an orderIndependent
loop. Calling it shortly after function resolution returns true on a simple
stream-like loop. It is currently just printing a message when it finds
suitable loops for GPU, but we can add a field to ForLoop that is set for
GPU-capable loops that later transformations can key off of.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>