Update GPU suitable code to be slightly less prototype
Add a field to the ForLoop AST that tracks if a for loop is suitable for
running on GPU. Add methods to get and set the value of the new field.

Move the function that analyzes blocks to determine if they can run on a
GPU into its own file along with a pass to run it on all of the ForLoops in
the program.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>