Also determine oversubscription via number of locales on local node.
If the CHPL_RT_OVERSUBSCRIBED is not set on chapcs then libfabric will choose
the "verbs;ofi_rxm" provider which causes internal errors for reasons yet to be
determined. As an alternative, check the number of locales on the local node to
determine oversubscription.  This is more reliable than the
CHPL_RT_OVERSUBSCRIBED environment variable but isn't yet implemented on all
platforms so for now consider both.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>