Add checking of ForallStmts to GPU suitability checker.
Add a field to the ForallStmt AST that tracks if a forall loop is suitable for
running on GPU. Add methods to get and set the value of the new field.

Add a pass to the GPU suitability checker that walks over forall loops checking
them and setting the field in ForallStmt if they are suitable.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>