Improvements to Poi tracking
It's not enough to track the Poi in the TypedFnSignature because it can
be shared with signatures used by multiple pois that ultimately have
different bodies. So, add a poiScope to ResolvedExpression to track the
poi to use when computing resolvedFunction corresponding to the call.

Additionally, adds resolvedConcreteFunction which is simpler for
tests to call and probably one day useful for export functions.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>