Extend bounded coforall optimization to zippered loops
5018 and 6419 optimized bounded coforalls over ranges/domains/arrays
by reducing the number of times we manipulate the endCount. However, it
only did this for direct loop, and not zippered ones. The lack of this
optimization led to regressions from 17354, which changed some direct
coforalls to zippered ones. That PR reduced the amount of communication,
but the overhead from not triggering the bounded coforall optimization
outweighed that, so it was partially reverted in 17531.

This extends the optimization to trigger for zippered coforalls as well,
which should allow us to revert 17531.

Resolves private 1925
Resolves private 1850

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>