Add llvm handling for basic macros used by LAPACK tests
When we encounter the LAPACK_GLOBAL macro add it and its clang::MacroInfo to
the LayeredValueTable. Then when processing the macros that call LAPACK_GLOBAL,
look it up in the LVT and compute the call's result using the arguments at
the call site. Add the mapping from the outer macro's name to the call's
computed forwarding result to the LVT.

When looking up a name in the LVT, check if it has a forwarded name and if
it does use that instead.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>