Deprecate `use Mod except *;`, we want to enable specifying operators
`except *;` today means exclude everything in the module and only enable
qualified access.  However, `*` is also the multiplication operator and we would
like to enable listing operators in `only` and `except` lists, as well as for
import statements.  It would be confusing to switch the meaning of `except *`
from under a user so deprecate it for at least a release cycle.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>