Fix array reallocation under ugni
Arrays larger than 2 hugepages are allocated through the comm layer
interface under comm=ugni for NUMA affinity reasons. For simple
allocation/free this works well, but there were some issues with
reallocations that this PR fixes.

Previously when a comm allocated array was shrunk with realloc if it
shrunk to under 2 hugepages we would call the normal non-comm realloc on
it, which led to cross-allocator calls and undefined behavior. This
fixes that by preventing reallocs when crossing the normal <-> comm
allocated threshold. Instead, call sites are required to do a full
alloc+copy+free in this case.

The cleaner way to do this would be to have clients track whether their
memory was comm allocated or not and do away with the threshold checks
when reallocating and freeing, but that will be more invasive.

This fixes the arrayRealloc test under ugni.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>