Move code to check if a block is ok for GPU to the end of deadCodeElimination
The GPU outlining work is currently being done at the end of
deadCodeElimination, so move the viability checking function there. Make
updates to allow it to handle the much more complicated AST at this later
point in the compiler.

Currently we believe this is too permissive as it allows calling functions
that are themselves safe, but we don't want to handle any function calls at
this time.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>