Improve error message when a CHPL_LLVM=none build is used with CHPL_LLVM!=none
I keep hitting this error and finding the current message difficult
to understand, particularly putting myself in the shoes of someone
who built Chapel with the quickstart build and then is trying to
use it from a new shell that may not set CHPL_LLVM=none.  This tries
to make it much more explicit as proposed in #18073.

Resolves #18073.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>