Use gcc 10 for Arkouda XC release testing
This system version was just upgraded to 11 and we get some warnings
from 11 that are turned into build failures with `WARNINGS=1`, which
throws `-Werror`. These warnings were fixed in 18178, but this is
testing against Chapel 1.24.1 so those changes aren't available. Just
pin to gcc 10 until this testing moves on to 1.25.0.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>