Fix an issue involving 32-bit floating point math in the LLVM backend
Fix an issue that was leading to an internal error with the LLVM backend that
caused it to print "DestTy too big for FPTrunc" and fail a consistency check.

When code generating an operation where the lhs was real(32) and the rhs was
real(64) we were generating FPTrunc(lhs, rhs.type). We instead need
FPExt(lhs, rhs.type).

Added a test derived from the code in issue #15817 that reproduces the error
prior to this fix.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>