Stop ignoring declared types when initializing expressions are tuples
This fixes a bug in our code base that Shreyas ran into in which if a
variable's initializer is a tuple:

   var r: T = (t1, t2);

then the type of the variable 'T' is completely ignored and 'r' is
treated as a tuple, as though it had been declared:

   var r = (t1, t2);

It turns out that this was because when resolution found any
declaration initialized with a tuple, it used a tuple initCopy()
routine to initialize the variable.  Here, I only make this initCopy()
rule fire if both the rhs and lhs are tuple types.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>