Insight: First try without actual args, then with
This escaped me before.  If we look at the resolution of foo(i) and we
only have the single foo(x: int) overload, we first try to resolve
'foo', and then add additional arguments only if we can't resolve it.
That suggests to me that we should simply bail out early if we have a
candidate vector with both paren-ful and paren-less functions in it.
Will try that next.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>