Upgrade to GASNet-EX 2021.8.0 snapshot
The next official GASNet release will be 2021.9.0, but that comes out
after our 1.25.0 release and there have been some some performance
improvements on InfiniBand that we want to include in our release.
Upgrade to a snapshot of the "stable" branch that the GASNet team has
added 2021.8.0 version markers to.

This improves performance on InfiniBand by serializing polling on
completion queues (CQs). Previously, we we getting high contention and
false-sharing on the underlying locks that verbs uses to protect CQs. CQ
polling is now protected by a cacheline aligned try-lock in gasnet,
which reduces contention and resolves the false-sharing issue. For more
info see https://bitbucket.org/berkeleylab/gasnet/pull-requests/468.

This serialization improves performance of concurrent communication. On
the cascade lake nodes of 16-node-cs-hdr we see large improvements for
RA-on, minor improvements for RA-rmo, as well as modest improvements for
aggregated indexgather and NPB-FT. On the rome nodes of 16-node-cs-hdr
performance for RA and NPB didn't change much, but indexgather improved
more.

16-node-cs-hdr with 48-core CascadeLake CPUs:

| config | RA-on       | RA-rmo      | indexgather    | NPB-FT     |
| ------ | ----------: | ----------: | -------------: | ---------- |
| before | 0.0018 GUPS | 0.0010 GUPS | 1075 MB/s/node | 470 GFLOPS |
| after  | 0.0080 GUPS | 0.0011 GUPS | 1150 MB/s/node | 490 GFLOPS |

16-node-cs-hdr with 128-core Rome CPUs

| config | indexgather    |
| ------ | -------------: |
| before | 1600 MB/s/node |
| after  | 1875 MB/s/node |

This also extends Cray PMI support to work on HPE Cray EX systems. See
the ChangeLog for additional details and other improvements.

Resolves private 2040
Helps 14438

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>