Automatically detect oversubscription
The function chpl_set_num_locales_on_node is used to set the number of
locales on the local node as determined by the communication system.
This function also determines if the node is oversubscribed based on
the number of locales, although this can be overidden by the
CHPL_RT_OVERSUBSCRIBED environment variable. The functions
chpl_get_num_locales_on_node and chpl_get_oversubscribed return the
number of locales on the local node and whether or not the node is
oversubscribed, respectively. This functionality is currently not
implemented on gasnet; we are waiting to upgrade to the EX API to do
so.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>