Deprecate range.ident()
We've talked about replacing this with '===' in the long-term, and
this would not be hard to do in the short-term, but also conjectured
that it is probably virtually unused/unknown to users, so proposed
also just deprecating it and seeing if anyone cried.  Even in our code
base, it seems we were just using it in one place, which is fewer than
I expected, though similarly miniscule.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>