Future for #18218, internal error, loop over findfiles() calls throwing function
I was unable to reproduce this error writing my own recursive iterator
in place of findfiles().  (It's really walkdirs() that findfiles()
calls that's recursive.)

I was concerned that a change to these iterators in the FileSystem
module might inadvertantly change the behavior of this test.  So I
added snapshots of findfiles and walkdirs() to the test for it to
call.

---
Signed-off-by: Paul Cassella <gitgitgit@manetheren.bigw.org>