Fix the user identifier of a paren-less and bracket-less function
When the parser+lexer is gobbling up text to store as the user's
function header, it has not been handling this case properly:

         proc parenLessFn return expr;

Because it's been triggering off of the presence of curly brackets as
the signal to stop gobbling code.  This makes a simple change to the
lexer to also stop if it sees a return statement (the one statement
that we support for functions without curly brackets at present).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>