Fix a reallocation bug when using a dynamic heap
For ugni we support a dynamically extendable heap. This functionality is
implemented using jemalloc's chunk hooks where we provide a hook to
acquire more memory from the "system". The alloc chunk hook takes `void*
chunk` and the man page says that "If chunk is not NULL, the returned
pointer must be chunk on success or NULL on error." We returned NULL if
the chunk didn't match for our fixed heap implementation, but this check
was missing from the dynamic heap implementation. This code path is only
taken when reallocating code where jemalloc is trying to grow in place
by grabbing an adjacent chunk and merge with an old one. Fix the code to
return `NULL` to indicate we aren't able to grow in place. It's possible
our allocation could happen to get the right address for serial
reallocations, but that's unlikely in the common case so we don't want
to waste time allocating/freeing.

This was causing sporadic segfaults for a user code detailed in
Cray/chapel-private 2394. We don't have a smaller reproducer for the
user code yet, but we're working on that in Cray/chapel-private 2427

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>