Use default timeouts of Cray CS testing
We used a higher timeout when ofi perf testing was originally added
because MiniMD and task spawning tests were slow. Looking at the most
recent nightly logs MiniMD takes under 2 minutes and task spawning is
under a minute. MiniMD is the longest executing test so I think it's
safe to use the default 5 minute timeout now.

Also remove regular cray cs timeout, which was added when we were having
problems after moving to an EDR machine that ended up being unstable.
We moved to a different HDR machine that has not had the same issues so
the timeouts are not needed.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>