Refine setting FI_MR_ALLOCATED in provider hints
Previously, FI_MR_ALLOCATED would not be set in the provider hints if the node
was oversubscribed, determined either automatically or via the
CHPL_RT_OVERSUBSCRIBED environment variable. This was done to avoid running
out of memory when allocating the fixed heap. However, the heap size is now
scaled by the number of locales on the local node, so it's ok to set
FI_MR_ALLOCATED in that case, but it should still not be set if
CHPL_RT_OVERSUBSCRIBED is set. In that situation the node might be
oversubscribed because it is running multiple applications.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>